Improve generate.py file
Add function that handles the API call and has some intelligent error
handling around the API call to ensure that we have good data, and if we
dont, we try another API call a set period of time afterwards.

Output many more columns to our output dataframe.

We now include every address-stop pair that we get from pruning (based
off of n_stops), but we output an additional column that reports which
address-stop pair is the shortest walking distance based on the google
distance matrix api call.

Spherical distance calculation now outputs in km to match the distance
output by the api.