Add Google API Wrapper class
Add a class to sanitize and manage calls to the google API. This is
intended to be easily replaced with another API should we choose to
use one.

Add API Wrapper tests. Tests are incomplete, as many calls to the
file system and the API still need to be mocked out.